server/requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
defusedxml==0.6.0
docx2txt==0.8
Flask==1.1.2
Flask-Cors==3.0.10
google==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
IMDbPY==2020.9.25
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
loguru==0.5.3
lxml==4.6.2
MarkupSafe==1.1.1
nltk==3.5
pycodestyle==2.6.0
PyExecJS==1.5.1
pygoogletranslation==2.0.4
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2020.5
regex==2020.11.13
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
toml==0.10.2
tqdm==4.56.0
twilio==6.51.0
Unidecode==1.1.2
urllib3==1.25.9
Werkzeug==1.0.1
win32-setctime==1.0.3